requirements.txt
Flask>=2.0
mysql-connector-python>=8.0
qrcode>=7.3
Pillow>=9.0
python-dotenv>=1.0
requests>=2.0
